+++ packages/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @forge-ts/api
 
+## 0.6.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @forge-ts/core@0.6.2
+
 ## 0.6.1
 
 ### Patch Changes